refactor(polymarket): split execution module by responsibility
Refactor the Polymarket execution client into focused modules for lifecycle, orders, cancellations, reports, and responses.

This change only reorganizes code distribution and keeps existing behavior intact. No new functionality is added and no execution paths are intentionally changed.

Closes #4257
